Refactor spell casting into dedicated hook
